moved stats page outside of course folder
